backend/requirements.txt
# ============================
# Core Python & Utilities
# ============================
numpy
opencv-python
Pillow
tqdm

# ============================
# PyTorch (Action + Anomaly + Explainability)
# ============================
torch
torchvision
torchaudio

# ============================
# Machine Learning Utilities
# ============================
scikit-learn

# ============================
# Explainability & Visualization
# ============================
matplotlib

# ============================
# NLP / Captioning (COCO-trained BLIP)
# ============================
transformers
sentencepiece
accelerate

# ============================
# JSON / Logging (standard, but safe)
# ============================
ujson
